fix: gofmt alignment and move e2e_ebpf helpers behind build tag
Move e2eImage, e2ePullPolicy, applyManifest to ebpf_test.go (which
has the e2e_ebpf build tag) so the linter doesn't flag them as unused
when running without the tag.
